fix: add kernel compression

+++ nixos/kernel/sdm845/default.nix
@@ -14,7 +14,6 @@ let
   );
 in
 {
-
   build =
     (pkgs.mobile-nixos.kernel-builder {
       inherit (configure) src;
@@ -23,6 +22,7 @@ in
         python3
         zstd
         kmod
+        gzip
       ];
 
       isModular = true;
@@ -30,27 +30,38 @@ in
       modDirVersion = "${configure.version}${configure.passthru.localVer}";
       makeImageDtbWith = "qcom/sdm845-oneplus-fajita.dtb";
       isCompressed = "gz";
-      installTargets = [ "modules_install" ];
+      installTargets = [
+        "zinstall"
+        "modules_install"
+      ];
     }).overrideAttrs
       (attrs: {
         passthru = attrs.passthru // {
           inherit kconfigToNix configure;
         };
 
-        installTargets = [ "modules_install" ];
+        installTargets = [
+          "zinstall"
+          "modules_install"
+        ];
         installFlags = [
           "INSTALL_MOD_PATH=$(out)"
           "INSTALL_PATH=$(out)"
+          "KBUILD_IMAGE=arch/arm64/boot/Image.gz"
         ];
 
         postInstall = ''
           mkdir -p $out
 
-          cp -v arch/arm64/boot/Image.gz $out/Image.gz
-          ln -sv Image.gz "$out/vmlinuz" || true
+          if [ -f arch/arm64/boot/Image.gz ]; then
+           cp -v arch/arm64/boot/Image.gz $out/Image.gz
+          elif [ -f arch/arm64/boot/Image ]; then
+            gzip -c arch/arm64/boot/Image > $out/Image.gz
+          fi
 
+          ln -sv Image.gz "$out/vmlinuz" || true
           cp .config $out/config-${configure.version}
-          depmod -b $out -F System.map "${configure.version}"
+          depmod -b $out -F System.map "${configure.version}${configure.passthru.localVer}"
         '';
 
         configurePhase = ''
